// src/app/models/shopping-cart.model.ts
export class ShoppingCart {
    _id: string = '';
    categoryId: string = '';
    productId: string = '';
    name: string = '';
    img: string = '';
    description: string = '';
    quantity: number = 0;
    price: number = 0;
    size: string = '';
    username: string = '';
    status: number = 0;
}